Fix simple remctl command array termination
The remctl simple call expects `command' to be a NULL-terminated
array of NUL-terminated strings. We had a subtle error in our
construction of the C array we passed the remctl library call
that resulted in an improper free() if the command had an odd
number of entries of length 3 or greater.