First pass at the `complex' interface
The general idea is that you create a New remctl instance, and
then Open a connection to a remctld host. Calling Execute
executes the command, and you listen on the instance channels
for Output, Error, Status and Done communication. Internally,
remctld output is gathered in a goroutine and send back over
the appropriate channel.